app2/logging-service/requirements.txt
fastapi
uvicorn
pydantic
hazelcast-python-client